requirements.txt

snowflake-connector-python
snowflake-snowpark-python
requests
pandas

